+++ src/Model/ISpellingSpecification.ts
@@ -1,0 +1,7 @@
+import ISpelling from "./ISpelling"
+
+export default interface ISpellingSpecification {
+	id: string
+	speller: new (id: any, options?: any) => ISpelling
+	options?: any
+}